requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
fastapi==0.129.0
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai==2.21.0
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pgvector @ git+https://github.com/pgvector/pgvector-python.git@1a72b7571adf3325174b383aca85bfb3a5b925fa
pillow==12.1.1
pip-review==1.3.0
pipdeptree==2.31.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.19
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.3
setuptools==82.0.0
shap==0.50.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.52.1
style==1.1.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update==0.0.1
urllib3==2.6.3
uvicorn==0.41.0
wheel==0.46.3
xgboost==3.2.0
